Working action bar
I fixed up all the links in the action bar to specify the location in
PhantomRunner activity, where you can put the intents and such. I
removed the button in the main and wired up the map class to the run
button in the action bar. I put a text field in the middle of the main
to show the button clicked. I'm plaaning on creating a view so when the
timer is clicked that pops up allowing the user to enter the time and
distance then launch the intent, with the distance and time in the
string i pass to the map class